chore; move signature and address verification to modal validator
